avance pagina admin, petición

+++ frontEnd/src/app/administrador/pages/productos/productos.component.ts
@@ -13,6 +13,7 @@ import { Products } from '../../../interfaces/products';
 
 
 
+
 @Component({
   selector: 'app-productos',
   standalone: true,
@@ -72,15 +73,24 @@ export class ProductosComponent {
 
   borrarProducto(id: string) {
     console.log("producto que se borrara tiene el id =", id);
-    this.producto.deleteProducts(id).subscribe((res: any) => {
-      try {
-        if (res) {
-          console.log("res", res)
-        }
-      } catch (error) { console.log(error) }
-    })
-
+    let isConfirm: boolean = false 
+    isConfirm = confirm("Esta seguro que quiere eliminar este producto");
+    if (isConfirm){
+      this.producto.deleteProducts(id).subscribe((res: any) => {
+        try {
+          if (res) {
+            console.log("res", res);
+            alert("Se elimino correctamente")
+            this.obtenerProductos()
+          }
+        } catch (error) { console.log(error) }
+      })
+    }
   }
+
+ 
+
+// logica para crear producto
 
   crearProductos() {
 
@@ -100,8 +110,6 @@ export class ProductosComponent {
         cantidad: this.cantidad,
         precio: this.precio
       }
-
-
       this.producto.postProducts(nuevoProducto).subscribe((res: any) => {
         try {
           if (res) {
@@ -110,9 +118,6 @@ export class ProductosComponent {
         } catch (error) { alert(error) }
       })
     }
-
-
-
   }
 
   identificarId(id: string) {
